T.4.3 second solution

+++ Module_4/T3-overload_operators/src/geomvector.cpp
@@ -19,7 +19,7 @@ std::ostream &operator<<(std::ostream& out, const GeomVector& a) {
 double GeomVector::Length() const {
     return sqrt(pow(x_, 2) + pow(y_, 2) + pow(z_, 2));
 }
-GeomVector operator/(const GeomVector& b, int a){
+GeomVector operator/(const GeomVector& b, double a){
     return GeomVector(b.x_/a, b.y_/a, b.z_/a);
 }
 bool operator>(const GeomVector& a, const GeomVector& b){
